+++ src/widgets/settingspages/ImageUploaderPage.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "widgets/settingspages/SettingsPage.hpp"
+
+namespace chatterino {
+
+class ImagePtrItemDelegate;
+
+class ImageUploaderPage : public SettingsPage
+{
+public:
+    ImageUploaderPage();
+};
+
+}  // namespace chatterino